+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Update to stripes v6. Refs UICHKIN-226.
 * Fix incorrect ref assignment. Refs UICHKIN-205.
 * Increment `@folio/stripes-cli` to `v2`. Refs UICHKIN-231.
+* Show confirmation modal when check in an item with one of the new statuses (Long missing, In process (non-requestable), Unavailable, Unknown). Refs UICHKIN-120.
 
 ## [4.0.0] (https://github.com/folio-org/ui-checkin/tree/v4.0.0) (2020-10-08)
 [Full Changelog](https://github.com/folio-org/ui-checkin/compare/v3.0.0...v4.0.0)